for some time delete frontend folder, add goals controller and upgrade spendings and earnings by adding dates and categories

+++ server/src/pkg/controllers/spending_controller.go
@@ -1,18 +1,18 @@
 package controllers
 
 import (
+	"time"
+
 	"github.com/gin-gonic/gin"
 	"github.com/szymon676/finance-tracker/server/src/pkg/database"
 	"github.com/szymon676/finance-tracker/server/src/pkg/models"
 )
 
 func HandleGetSpendings(c *gin.Context) {
-
 	var spendings []models.Spending
-
 	database.DB.Find(&spendings)
 	c.JSON(200, gin.H{
-		"spendings:": spendings,
+		"spendings": spendings,
 	})
 }
 
@@ -21,35 +21,49 @@ func HandleNewSpending(c *gin.Context) {
 	var data models.BindSpending
 
 	if err := c.BindJSON(&data); err != nil {
+		c.Error(err)
 		c.String(400, "Error binding in spending")
+		return
 	}
 
 	//checking if user is passing good data
-	if data.Title == "" || data.Price == "" || data.Description == "" {
+	if data.Ammount == 0 || data.Category == "" {
 		c.JSON(400, gin.H{
 			"status": "failed, please fill all fields",
 		})
-
-	} else {
-
-		income := models.Spending{Title: data.Title, Price: data.Price, Description: data.Description}
-		result := database.DB.Create(&income)
-
-		//handling errors when creating user is failed
-		if result.Error != nil {
-			c.JSON(400, gin.H{
-				"status": "failed",
-			})
-			return
-
-		} else {
-			// creating new user if result is successful
-			c.JSON(202, gin.H{
-				"status": "success",
-			})
-
-		}
+		return
 	}
+
+	layout := "02.01.2006"
+	date, err := time.Parse(layout, data.Date)
+
+	if err != nil {
+		c.String(400, "error in date format")
+		return
+	}
+
+	if date.After(time.Now()) {
+		c.String(400, "date cannot be in the future")
+		return
+	}
+
+	spending := models.Spending{
+		Ammount:  data.Ammount,
+		Category: data.Category,
+		Date:     date,
+	}
+
+	result := database.DB.Create(&spending)
+	if result.Error != nil {
+		c.JSON(400, gin.H{
+			"status": "failed",
+		})
+		return
+	}
+
+	c.JSON(202, gin.H{
+		"status": "success",
+	})
 }
 
 func HandleDeleteSpending(c *gin.Context) {
@@ -57,12 +71,24 @@ func HandleDeleteSpending(c *gin.Context) {
 	var spending models.Spending
 	id := c.Param("id")
 
-	database.DB.First(&spending, id)
+	result := database.DB.First(&spending, id)
 
-	if spending.Title == "" {
+	if result.RowsAffected == 0 {
 		c.String(404, "spending not found delete failed")
+		return
 	}
+
 	database.DB.Delete(&spending, id)
 	c.String(202, "success")
+}
 
+func HandleGetSpendingsByCategory(c *gin.Context) {
+
+	var spendings []models.Spending
+	category := c.Param("category")
+	database.DB.Where("category = ?", category).Find(&spendings)
+
+	c.JSON(200, gin.H{
+		"spendings": spendings,
+	})
 }
